First Gerbers Generated...
The gerbers check out fine in GERBV, however they are incompatible with
OSHParks service.  There are two drill files (PTH and Non-PTH) whereas
OSHPark expects only one.  Additionally, the oval drills for the 3.5mm
stereo jack pins is not registering as oval but circular which would not
allow the component to mount successfully. This will need to be
rectified with OSHPark Support tomorrow.